Move math hint tests to math_utils
